fix: move schema template helper to surveyutils
