+++ pkg/cmd/step/verify/test_data/preinstall/terraform_kaniko/jx-requirements.yml
@@ -1,9 +1,35 @@
-kaniko: true
-terraform: true
-secretStorage: local
 cluster:
   clusterName: my-cluster-name
-  provider: gke
-  project: foo
-  zone: myzone
   environmentGitOwner: myorg
+  gitKind: github
+  gitName: github
+  gitServer: https://github.com
+  namespace: jx
+  project: foo
+  provider: gke
+  zone: myzone
+ingress:
+  domain: ""
+  externalDNS: false
+  namespaceSubDomain: -jx.
+  tls:
+    email: ""
+    enabled: false
+    production: false
+kaniko: true
+secretStorage: local
+storage:
+  logs:
+    enabled: false
+    url: ""
+  reports:
+    enabled: false
+    url: ""
+  repository:
+    enabled: false
+    url: ""
+terraform: true
+versionStream:
+  ref: v1.0.3
+  url: https://github.com/jenkins-x/jenkins-x-versions.git
+webhook: prow
