+++ pkg/surveyutils/test_data/template/jx-requirements.yml
@@ -1,0 +1,7 @@
+cluster:
+  clusterName: foo
+  gitKind: gitlab
+  gitServer: https://gitlab.com
+  namespace: jx
+  provider: gke
+secretStorage: vault
